+++ CHANGELOG.markdown
@@ -1,6 +1,10 @@
 CHANGELOG for the Transifex API Package
 ===============
 
+* 1.2.0 (2015-07-13)
+
+ * Add `Transifex::get()` to fetch API objects
+
 * 1.1.0 (2015-07-12)
 
  * Deprecated `TransifexObject::processResponse()`, 2.0 will return the full `Joomla\Http\Response` object instead of processing the response internally